ecommerce_drf/requirements.txt
asgiref==3.5.0
autopep8==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.12.0
django-js-asset==1.2.2
django-modeltranslation==0.15
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
gprof2dot==2019.11.30
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
psycopg2==2.8.6
pycodestyle==2.5.0
pycparser==2.21
Pygments==2.6.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.1
python-gettext==4.0
python3-openid==3.2.0
pytils==0.3
pytz==2019.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
